# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Chapter:
    properties:
      chapterNumber:
        type: string
      createdAt:
        type: string
      id:
        type: string
      mangaID:
        type: string
      pages:
        items:
          type: string
        type: array
      title:
        description: Optional
        type: string
      updatedAt:
        type: string
    type: object
...
  handler.createChapterRequest:
    properties:
      chapter_number:
        maxLength: 20
        type: string
      pages:
        description: Initially, pages might be empty before upload
        items:
          type: string
        type: array
      title:
        maxLength: 255
        type: string
    required:
    - chapter_number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a production-ready HTTP API in Go for a manga reading platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Manga-Dex-Style API
  version: "1.0"
paths:
...
  /chapters/{id}:
    delete:
      description: Deletes a specific chapter. Requires 'chapters:manage' permission.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a chapter
      tags:
      - Chapters
    get:
      description: Retrieves details for a single chapter.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Chapter'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single chapter by ID
      tags:
      - Chapters
    put:
      consumes:
      - application/json
      description: Updates the details of a specific chapter. Requires 'chapters:manage'
        permission.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      - description: Chapter Update Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createChapterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Chapter'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a chapter
      tags:
      - Chapters
...
  /chapters/{id}/pages:
    post:
      consumes:
      - multipart/form-data
      description: Uploads one or more image files for a chapter. Requires 'chapters:manage'
        permission.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      - description: Image files for the chapter pages. Can be sent multiple times.
        in: formData
        name: pages
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload chapter pages
      tags:
      - Chapters
...
  /manga/{manga_id}/chapters:
    get:
      description: Retrieves a paginated list of chapters for a specific manga.
      parameters:
      - description: Manga ID
        in: path
        name: manga_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Chapter'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List chapters for a manga
      tags:
      - Chapters
    post:
      consumes:
      - application/json
      description: Adds a new chapter to a specific manga. Requires 'chapters:manage'
        permission.
      parameters:
      - description: Manga ID
        in: path
        name: manga_id
        required: true
        type: string
      - description: Chapter Creation Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createChapterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Chapter'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new chapter
      tags:
      - Chapters
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"